requirements_new.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.0
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==1.26.4
openai==2.3.0
openpyxl==3.1.2
pandas==2.2.0
pillow==11.3.0
pydantic==2.12.1
pydantic_core==2.41.3
Pygments==2.19.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
